+++ packages/styleguide/src/lib/Molecules/Tips/ToolTip/ToolTip.stories.tsx
@@ -1,8 +1,16 @@
-import { FillButton, FlexBox, IconButton, ToolTip } from '@codecademy/gamut';
+import {
+  FillButton,
+  FlexBox,
+  IconButton,
+  StrokeButton,
+  ToolTip,
+} from '@codecademy/gamut';
 import {
   ArrowRightIcon,
+  DeleteIcon,
   SmileyStarEyesIcon,
   SparkleIcon,
+  StudyBookIcon,
 } from '@codecademy/gamut-icons';
 import type { Meta, StoryObj } from '@storybook/react';
 
@@ -16,7 +24,7 @@ type Story = StoryObj<typeof ToolTip>;
 
 export const Default: Story = {
   render: () => (
-    <FlexBox center py={64} m={24}>
+    <FlexBox center m={24} py={64}>
       <ToolTip id="fill-id" info="Tooltip">
         <FillButton aria-describedby="fill-id" aria-disabled icon={SparkleIcon}>
           Click me
@@ -30,16 +38,15 @@ export const WithIconButton: Story = {
   render: () => (
     <FlexBox justifyContent="space-around" m={24} width="95%">
       <IconButton
-        tip="Beautify your code"
         icon={SparkleIcon}
+        tip="Beautify your code"
         tipProps={{ alignment: 'bottom-center' }}
       />
       <IconButton
-        aria-label="Next Prompt"
-        tip="Next Prompt"
         icon={ArrowRightIcon}
+        tip="Next Prompt"
+        tipProps={{ alignment: 'bottom-center' }}
         variant="secondary"
-        tipProps={{ alignment: 'bottom-center', hideAriaToolTip: true }}
       />
     </FlexBox>
   ),
@@ -47,11 +54,49 @@ export const WithIconButton: Story = {
 
 export const Floating: Story = {
   render: () => (
-    <IconButton
-      aria-label="Wonder"
-      tip="Wonder at the majesty of the universe"
-      icon={SmileyStarEyesIcon}
-      tipProps={{ alignment: 'bottom-center', placement: 'floating' }}
-    />
+    <FlexBox center justifyContent="space-around" m={24}>
+      <IconButton
+        icon={SmileyStarEyesIcon}
+        tip="Wonder at the majesty of the universe"
+        tipProps={{ alignment: 'bottom-center', placement: 'floating' }}
+      />
+      <ToolTip
+        id="floating-ex"
+        info="Tooltip for a FillButton"
+        placement="floating"
+      >
+        <FillButton aria-describedby="floating-ex" icon={StudyBookIcon}>
+          Also floating
+        </FillButton>
+      </ToolTip>
+    </FlexBox>
+  ),
+};
+
+export const InteractiveElement: Story = {
+  render: () => (
+    <FlexBox center m={32}>
+      <ToolTip id="stroke-button-ex" info="And here's a tooltip">
+        <StrokeButton aria-describedby="stroke-button-ex">
+          I&apos;ve got my own text
+        </StrokeButton>
+      </ToolTip>
+    </FlexBox>
+  ),
+};
+
+export const Disabled: Story = {
+  render: () => (
+    <FlexBox center m={24}>
+      <ToolTip id="disabled-ex" info="Tooltip still shows" placement="floating">
+        <FillButton
+          aria-describedby="disabled-ex"
+          aria-disabled
+          icon={DeleteIcon}
+        >
+          Using aria-disabled
+        </FillButton>
+      </ToolTip>
+    </FlexBox>
   ),
 };